Add "24th hour of the day" functionality.
We are removing the STATIC_DAY and instead calling `moment(str,
"HH:mm")` whenver we want to parse a string into a moment object. This
means that if we have an array `["8:00", "9:00"]`, we may end up with
the first moment call being parsed at one day, and the next moment call
at the next day. The odds of running into this issue are small. The
gains we get here is that it allow us to properly add "24 hour"
functionality to our calendar. This is a good tradeoff.