Make two-column layout more flexible.
The idea here is that we are going to display all of the times in two
columns. Therefore, let's split times into two groups, and iterate over
these two groups.

We could make this even more flexible by chunking times into n groups of
x, where x is the maximum number of time objects we want to show per
column. I don't need this flexibility yet, so I am not adding it.